Solving Recursively
Everytime a number is filled in the solve method is called again to restart the process and look for a 0. This was to stop the error of overwriting a changed number and also a bug where 1's were filled in where 0 was found in incorrect places. (This is testing for the first 3 cells in the grid).